+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Improvements
 
+- Improve session replay frame presentation timing calculations (#5133)
 - Use wider compatible video encoding options for Session Replay (#5134)
 
 ## 8.49.1